Moved the ListModeData_dummy to a new header
